+++ Iso20022Library.Domain/Common/Enums/MessageType.cs
@@ -25,5 +25,6 @@ public enum MessageType
     Pain00700108,
     Pain00700109,
     Pain00700110,
-    Pacs00200111
+    Pacs00200111,
+    Pacs00300108
 }